+++ 1brc-single-thread/src/hash.rs
