+++ 1brc-single-thread/src/iter/mod.rs
@@ -1,2 +1,5 @@
-pub mod line_chunks;
-pub mod raw_records;
+mod line_chunks;
+mod raw_records;
+
+pub use line_chunks::IterLineChunks;
+pub use raw_records::IterRawRecords;
